# docs/swagger.yaml
basePath: /
definitions:
...
  dto.APIResponse:
    properties:
      data:
        description: Dữ liệu trả về (nếu có)
      error:
        description: Lỗi (nếu có)
        example: ""
        type: string
      message:
        description: Thông báo
        example: Thao tác thành công
        type: string
      success:
        description: Request thành công hay thất bại
        example: true
        type: boolean
    type: object
...
  dto.ChangePasswordRequest:
    properties:
      new_password:
        example: newpass123
        maxLength: 100
        minLength: 6
        type: string
      old_password:
        example: oldpass123
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        maxLength: 100
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - manager
        - staff
        - customer
        example: staff
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      is_active:
        example: true
        type: boolean
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: 24/01/2026 10:00
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: uuid-123
        type: string
      is_active:
        example: true
        type: boolean
      is_email_verified:
        example: true
        type: boolean
      role:
        example: staff
        type: string
      updated_at:
        example: 24/01/2026 10:30
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant.local
    name: API Support
  description: API quản lý menu nhà hàng - Xây dựng theo Clean Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Menu API
  version: 2.0.0
paths:
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Lấy danh sách tất cả users (Manager+)
      parameters:
      - description: Filter by role
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Tạo user mới (Manager+)
      parameters:
      - description: Thông tin user mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Tạo user mới
      tags:
      - Users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Vô hiệu hóa user (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Vô hiệu hóa user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Lấy thông tin user theo ID (Manager+)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Lấy user theo ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin user (Manager+)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật user
      tags:
      - Users
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Lấy thông tin của user đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin user hiện tại
      tags:
      - Users
  /api/users/me/password:
    put:
      consumes:
      - application/json
      description: Đổi mật khẩu của user đang đăng nhập
      parameters:
      - description: Thông tin đổi mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Đổi mật khẩu
      tags:
      - Users
...
schemes:
- http
swagger: "2.0"